refactor: move VoteOnBallot to utils.go
